+++ test/support/test_header_controller.ex
@@ -1,6 +1,6 @@
 defmodule TestHeaderController do
   @moduledoc false
-  use PhoenixSpec.Controller, formats: [:json]
+  use PhoenixSpectral.Controller, formats: [:json]
 
   @type required_headers :: %{required(:"x-user-id") => String.t()}
   @type optional_headers :: %{optional(:"x-trace-id") => String.t()}